#493 Split company dropdown options in groups
