+++ src/i18n/translations/zh-Hans.json
@@ -11,6 +11,7 @@
     "Close": "关闭",
     "Company": "公司",
     "Company code": "公司代码",
+    "Core companies": "主要公司",
     "Follow the instructions below to open an Issue": "按下方的指引打开 Issue",
     "Import from": "导入自",
     "Justification for major update of": "重大更新的理由：",
@@ -18,6 +19,8 @@
     "Line code": "线路代码",
     "Local": "本地",
     "Major update": "重大更新",
+    "New": "添加",
+    "Other companies": "其他公司",
     "Please provide suitable source and justification.": "请您提供适当的来源和理由。",
     "Railway company": "铁路公司",
     "Remove this line": "删除此线路",
