+++ src/i18n/translations/zh-Hant.json
@@ -11,6 +11,7 @@
     "Close": "關閉",
     "Company": "公司",
     "Company code": "公司代碼",
+    "Core companies": "主要公司",
     "Follow the instructions below to open an Issue": "按下方的指引開啟 Issue",
     "Import from": "讀入自",
     "Justification for major update of": "重大更新的理由：",
@@ -18,6 +19,8 @@
     "Line code": "路線代碼",
     "Local": "本機",
     "Major update": "重大更新",
+    "New": "新增",
+    "Other companies": "其他公司",
     "Please provide suitable source and justification.": "請你提供適當的來源及理由。",
     "Railway company": "鐵路公司",
     "Remove this line": "移除此路線",
